🧹 [code health improvement] refactor cost-control.ts into multiple files
